requirements.txt
# Runtime dependencies
Flask==2.2.3
SQLAlchemy==2.0.0
Flask-SQLAlchemy==3.0.2
psycopg2==2.9.5
python-dotenv==0.20.0

# Runtime tools
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.14.0
flake8==5.0.4
black==22.3.0

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==2.12.0
coverage==6.3.2
# codecov==2.1.12

# Utilities
httpie==3.2.1
sqlalchemy-json==0.5.0
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
sqlalchemy-utils==0.40.0
boto3==1.26.76
flask-restx==1.0.6
Flask-Migrate==4.0.4
PyMuPDF==1.21.1
pillow==9.4.0
pandas==1.5.3